safiretools: removes HamiltonianDirector and moves some functionality to HamiltonianBuilder
